Bugfix (Client): use proxy consistently
Since the legacy `load_configuration` returns a proxy as string, we can
not use it in place of the expected `dict` in `Client` constructor
(param `proxies`).

So, we just use `proxy` instead of `proxies`, and `proxy` is always an URL.

Session proxies for both HTTP and HTTPS are set to the same `proxy`, as
given by the user of `Client`, or defined in env, or in legacy config
file.

In addition to proxy changes, rename `url` to `endpoint`, also for
consistency.