Add feature (new config): failback to defaults for profile
As the absolute last resort when trying to decide which profile section
to load from the config file, failback to `[defaults]` section.

So, in summary: if profile keyword is undefined when calling
`load_config`, profile name is inferred from `DWAVE_PROFILE`
environment variable, and if that variable is not present,
`profile` key is looked-up in the `[defaults]` config section.
If `profile` is not defined under `[defaults]`, the first section
is used. If no other sections are defined besides `[defaults]`,
the `[defaults]` section is promoted to profile.