Define Future ordering as workaround for PriorityQueue bug
To handle somewhat unlikely case of polling tasks with
equal priority (double precision floating point timestamp).

Since we use `(timestamp, future)` for polling tasks in `PriorityQueue`,
if timestamps are equal, and futures have to be compared, enqueueing
will fail in an inconsistent state: the item *will* be added to queue, but
(internal) number of unfinished tasks will *not* be incremented. This
eventually leads to `ValueError` on `task_done()` after legitimate queue
`get`.