Set limit on problem submit batch size (closes #190)
Better solution would be to tally actual problem size when encoded
(which includes custom parameters, not only h/J), but a naive
implementation of that with FIFO problem queue would mutate submit
order. One solution would be to switch to priority queue for problems,
with problem sequence number as priority.