Update CLI tests after breaking changes in click 8.2
`mix_stderr` option has been removed between 8.1.8 and 8.2, without
any warning/deprecation period. The default is now the opposite of what
previously was the default (stdout/stderr is NOT mixed anymore;
previously is was mixed by default and we needed it to be separated
for some logging tests).

We can't simply require `click>=8.2` because it requires py310.
When we drop py39, we can upgrade click to 8.2 and simplify the test
logic.